Add RNA maturation process (#1299)
* Add back tRNA operons to list of transcription units

* Add separate table for mature RNAs in transcription sim_data

* Replace miscRNAs in complexation reactions with matching transcription units (#1298)

* Replace RNAs in complexation reactions with corresponding TUs

* Limit mapping to miscRNAs with single-gene TUs

* Remove compartment tags from tRNA flat files

* Add RnaMaturation process

* Fix all analysis scripts to be compatible with new listener outputs

+++ wholecell/containers/bulk_objects_container.py
@@ -327,7 +327,7 @@ class BulkObjectsContainer(object):
 			retrieval.  Otherwise modifying an output array (e.g. an array
 			return by .counts()) would modify the underlying data.
 		"""
-		return np.array([self._objectIndex[name] for name in names])
+		return np.array([self._objectIndex[name] for name in names], dtype=np.int64)
 
 
 	def __eq__(self, other):
